Remove expensive error formatting.
Query match errors are so common that the cost of formatting their string
values into the error text more than doubles the memory cost of the naive
benchmarks. It is likely no one ever sees these strings; if that turns out to
be wrong, we can probably do better by returning a structured error.